requirements.txt
--extra-index-url https://pytorch.org
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
anyio
archspec
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
audioread==3.0.1
auraloss==0.4.0
babel
beartype==0.19.0
beautifulsoup4
bleach
boltons
Brotli
cached-property
cachetools==5.5.2
certifi
cffi
charset-normalizer
click==8.2.1
cloudpickle==3.1.1
colorama
colorlog==6.9.0
comm
contourpy
cycler
Cython
dask==2025.5.1
dask-jobqueue==0.9.0
debugpy
decorator
defusedxml
distributed==2025.5.1
distro
einops==0.8.0
exceptiongroup
executing
fast-bss-eval
fastjsonschema
fastrlock==0.8.3
ffmpeg-python==0.2.0
filelock==3.13.1
fonttools
fqdn
frozendict
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gcsfs==2025.3.2
google-api-core==2.24.2
google-auth==2.39.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.70.0
h11
h2
h5py
hpack
httpcore
httpx
hydra-core @ git+https://github.com/facebookresearch/hydra.git@7c8fa4a7594fe0c333ce6180adecd50f6d1f43b9
hyperframe
idna
importlib_metadata
importlib_resources
ipykernel
ipython
ipywidgets
isoduration
jedi
Jinja2
joblib
json5
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kiwisolver
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.44.0
locket==1.0.0
Mako==1.3.8
Markdown==3.7
markdown-it-py
MarkupSafe
matplotlib==3.10.0
matplotlib-inline
mdurl
mir_eval==0.7
mistune
mpi4py==4.0.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
munkres==1.1.4
musdb==0.4.2
museval==0.4.1
mypy
mypy_extensions
nbclient
nbconvert
nbformat
nest_asyncio
networkx==3.3
notebook
notebook_shim
numba==0.61.2
numpy
oauthlib==3.2.2
omegaconf==2.4.0.dev3
opt_einsum==3.4.0
optuna==4.2.0
overrides
packaging
pandas
pandocfilters
parso
partd==1.4.2
pesq==0.0.4
pexpect
pickleshare
pillow
pkgutil_resolve_name
platformdirs
pluggy
pytest
pooch==1.8.2
progressbar2
prometheus_client
prompt_toolkit
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.28.3
psutil
ptyprocess
pure_eval
pyaml==25.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycparser
Pygments
pyparsing
pyroomacoustics==0.8.3
PySide6==6.8.1
PySocks
pystoi==0.4.1
python-dateutil
python-json-logger
python-utils
pytorch-lightning==2.5.0.post0
pytz
PyYAML
pyzmq
referencing
requests
requests-oauthlib==2.0.0
rfc3339_validator
rfc3986-validator
rich
rotary-embedding-torch==0.6.1
rpds-py
rsa==4.9.1
ruamel.yaml
ruamel.yaml.clib
ruff
schedulefree==1.4
scikit-learn
scipy
Send2Trash
setuptools<82
shiboken6==6.8.1
simplejson==3.20.1
six
sniffio
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve
soxr==0.5.0.post1
spafe==0.3.3
SQLAlchemy==2.0.37
stack_data
stempeg==0.2.3
sympy==1.13.1
tblib==3.1.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
terminado
threadpoolctl
tinycss2
tomli
toolz==1.0.0
torch
torchaudio
torchmetrics==1.6.1
torchvision
tornado
tqdm
traitlets
truststore
types-python-dateutil
typing_extensions
typing_utils
tzdata
unicodedata2
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension
yarl==1.18.3
zict==3.0.0
zipp
zstandard==0.23.0
